prevent null pointer in deleteAll if a 'null' client is passed

+++ src/test/java/io/chino/java/testutils/DeleteAll.java
@@ -17,7 +17,7 @@ import java.util.LinkedList;
 import java.util.List;
 
 public class DeleteAll {
-    
+
     /**
      * Delete all objects created during this test suite.
      */
@@ -194,8 +194,9 @@ public class DeleteAll {
         } else if (apiClient instanceof Auth) {
             // nothing to delete for Auth
         } else {
-            throw new UnsupportedOperationException("deleteAll(" + apiClient.getClass().getSimpleName() +
-                    ") is not supported.");
+            if (apiClient != null)
+                throw new UnsupportedOperationException("deleteAll(" + apiClient.getClass().getSimpleName() +
+                        ") is not supported.");
         }
     }
